requirements.txt
appnope==0.1.0
decorator==4.0.6
Django==1.9.1
django-braces==1.8.1
django-cors-headers==1.1.0
django-npm==0.1.4
django-oauth-toolkit==0.10.0
django-pipeline==1.6.4
django-pipeline-browserify==0.2.0
django-require==1.0.9
django-reset==0.2.0
django-rest-framework-social-oauth2==1.0.2
djangorestframework==3.3.2
futures==3.0.4
gnureadline==6.3.3
ipdb==0.8.1
ipython==4.0.3
ipython-genutils==0.1.0
oauthlib==1.0.3
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
ptyprocess==0.5
PyJWT==1.4.0
python-openid==2.2.5
python-social-auth==0.2.14
requests==2.9.1
requests-oauthlib==0.6.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.1.0
